+++ changelog.md
@@ -4,6 +4,7 @@
 - Allowed support for Eleventy >1.0.0.
 - Fixed crash occurring when compiling images with no parameters.
 - Fixed images in output of `parse()` having relative URL paths.
+- Fixed default template arguments not being parsed.
 
 ## 1.3.0
 *2021-05-16*